Rename and update README.md
We decided to provide a better GitHub experience by showing README.md as
default index page (as we cannot configure GitHub to show other files
for the index content).

In addition, we also updated the contents of README.md and added an
instruction on how to avoid conflicts with app's README.md file.